move fetching terraform dependencies out of inference and with other init logic
