+++ src/python/pants/backend/terraform/dependency_inference.py
@@ -4,22 +4,17 @@ from __future__ import annotations
 
 from dataclasses import dataclass
 from pathlib import PurePath
-from typing import Tuple
 
 from pants.backend.python.subsystems.python_tool_base import PythonToolRequirementsBase
 from pants.backend.python.target_types import EntryPoint
 from pants.backend.python.util_rules.pex import PexRequest, VenvPex, VenvPexProcess
 from pants.backend.python.util_rules.pex import rules as pex_rules
 from pants.backend.terraform.target_types import TerraformModuleSourcesField
-from pants.backend.terraform.tool import TerraformProcess
-from pants.backend.terraform.utils import terraform_arg, terraform_relpath
 from pants.base.glob_match_error_behavior import GlobMatchErrorBehavior
 from pants.base.specs import DirGlobSpec, RawSpecs
-from pants.core.util_rules.source_files import SourceFiles
 from pants.engine.fs import CreateDigest, Digest, FileContent
-from pants.engine.internals.native_engine import EMPTY_DIGEST, MergeDigests
-from pants.engine.internals.selectors import Get, MultiGet
-from pants.engine.process import FallibleProcessResult, Process, ProcessResult
+from pants.engine.internals.selectors import Get
+from pants.engine.process import Process, ProcessResult
 from pants.engine.rules import collect_rules, rule
 from pants.engine.target import (
     FieldSet,
@@ -142,64 +137,6 @@ async def infer_terraform_module_dependencies(
         tgt.address for tgt in candidate_targets if tgt.has_field(TerraformModuleSourcesField)
     ]
     return InferredDependencies(terraform_module_addresses)
-
-
-@dataclass(frozen=True)
-class GetTerraformDependenciesRequest:
-    source_files: SourceFiles
-    directories: Tuple[str, ...]
-    backend_config: SourceFiles
-
-    # Not initialising the backend means we won't access remote state. Useful for `validate`
-    initialise_backend: bool = False
-
-
-@dataclass(frozen=True)
-class TerraformDependencies:
-    fetched_deps: Tuple[Tuple[str, Digest], ...]
-
-
-@rule
-async def get_terraform_providers(
-    req: GetTerraformDependenciesRequest,
-) -> TerraformDependencies:
-    args = ["init"]
-    if req.backend_config.files:
-        args.append(
-            terraform_arg(
-                "-backend-config",
-                terraform_relpath(req.directories[0], req.backend_config.files[0]),
-            )
-        )
-        backend_digest = req.backend_config.snapshot.digest
-    else:
-        backend_digest = EMPTY_DIGEST
-
-    args.append(terraform_arg("-backend", str(req.initialise_backend)))
-
-    with_backend_config = await Get(
-        Digest, MergeDigests([req.source_files.snapshot.digest, backend_digest])
-    )
-
-    # TODO: Does this need to be a MultiGet? I think we will now always get one directory
-    fetched_deps = await MultiGet(
-        Get(
-            FallibleProcessResult,
-            TerraformProcess(
-                args=tuple(args),
-                input_digest=with_backend_config,
-                output_files=(".terraform.lock.hcl",),
-                output_directories=(".terraform",),
-                description="Run `terraform init` to fetch dependencies",
-                chdir=directory,
-            ),
-        )
-        for directory in req.directories
-    )
-
-    return TerraformDependencies(
-        tuple(zip(req.directories, (x.output_digest for x in fetched_deps)))
-    )
 
 
 def rules():
